+++ .gitignore
@@ -1,2 +1,3 @@
 config.json
 config.cfg
+helpers/__pycache__
